+++ apps/public-reference/pages/applications/start/choose-language.tsx
@@ -23,7 +23,7 @@ import React, { useContext, useEffect, useState } from "react"
 const loadListing = async (listingId, stateFunction, conductor, context) => {
   const response = await axios.get(process.env.listingServiceUrl)
   conductor.listing =
-    response.data.listings.find((listing) => listing.id == listingId) || response.data.listings[2] // FIXME: temporary fallback
+    response.data.listings.find((listing) => listing.id == listingId) || response.data.listings[0] // FIXME: temporary fallback
   const applicationConfig = retrieveApplicationConfig() // TODO: load from backend
   conductor.config = applicationConfig
   stateFunction(conductor.listing)
