+++ backend/core/ormconfig.test.ts
@@ -20,7 +20,7 @@ export = {
     // Needed to get a TS context on entity imports.
     // See
     // https://stackoverflow.com/questions/59435293/typeorm-entity-in-nestjs-cannot-use-import-statement-outside-a-module
-    join(__dirname, "src/entity/**", "*.ts"),
+    join(__dirname, "src/**", "*.entity.{js,ts}"),
   ],
   migrations: [join(__dirname, "src/migration/**", "*.ts")],
   subscribers: [join(__dirname, "src/subscriber/**", "*.ts")],
