+++ backend/core/src/applications/applications.module.ts
@@ -1,6 +1,6 @@
 import { Module } from "@nestjs/common"
 import { TypeOrmModule } from "@nestjs/typeorm"
-import { Application } from "../entity/application.entity"
+import { Application } from "./entities/application.entity"
 import { ApplicationsService } from "./applications.service"
 import { ApplicationsController } from "./applications.controller"
 import { AuthModule } from "../auth/auth.module"
@@ -8,9 +8,17 @@ import { CsvEncoder } from "../services/csv-encoder.service"
 import { CsvBuilder } from "../services/csv-builder.service"
 import { SharedModule } from "../shared/shared.module"
 import { ListingsModule } from "../listings/listings.module"
+import { Address } from "../shared/entities/address.entity"
+import { ApplicationData } from "./entities/application-data.entity"
+import { Applicant } from "./entities/applicant.entity"
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Application]), AuthModule, SharedModule, ListingsModule],
+  imports: [
+    TypeOrmModule.forFeature([Application, ApplicationData, Applicant, Address]),
+    AuthModule,
+    SharedModule,
+    ListingsModule,
+  ],
   providers: [ApplicationsService, CsvEncoder, CsvBuilder],
   exports: [ApplicationsService],
   controllers: [ApplicationsController],
