+++ backend/core/src/listings/listing.dto.ts
@@ -1,16 +1,16 @@
 import { Listing } from "../entity/listing.entity"
 import { ListingsResponseStatus } from "./listings.service"
 import { Exclude, Expose, Type } from "class-transformer"
-import { IsDefined, IsEnum, IsUUID, ValidateNested } from "class-validator"
+import { IsDefined, IsEnum, IsOptional, IsUUID, ValidateNested } from "class-validator"
 
 import { PreferenceDto } from "../preferences/preference.dto"
 import { AssetDto } from "../assets/asset.dto"
 import { ApplicationMethodDto } from "../application-methods/application-method.dto"
-import { UnitDto } from "../units/unit.dto"
 import { ApiHideProperty, ApiProperty, OmitType } from "@nestjs/swagger"
 import { ListingEventDto } from "../listing-events/listing-events.dto"
 import { IdDto } from "../lib/id.dto"
 import { PropertyDto } from "../property/property.dto"
+import { AddressUpdateDto } from "../shared/dto/address.dto"
 
 export class ListingDto extends OmitType(Listing, [
   "applicationMethods",
@@ -78,32 +78,57 @@ export class ListingCreateDto extends OmitType(ListingDto, [
   "property",
   "events",
   "applications",
+  "applicationAddress",
+  "applicationPickUpAddress",
+  "leasingAgentAddress",
 ] as const) {
   @Expose()
   @IsDefined()
   @ValidateNested({ each: true })
   @Type(() => IdDto)
   applicationMethods: IdDto[]
+
   @Expose()
   @IsDefined()
   @ValidateNested({ each: true })
   @Type(() => IdDto)
   assets: IdDto[]
+
   @Expose()
   @IsDefined()
   @ValidateNested({ each: true })
   @Type(() => IdDto)
   preferences: IdDto[]
+
   @Expose()
   @IsDefined()
   @ValidateNested()
   @Type(() => IdDto)
   property: IdDto
+
   @Expose()
   @IsDefined()
   @ValidateNested({ each: true })
   @Type(() => IdDto)
   events: IdDto[]
+
+  @Expose()
+  @IsOptional()
+  @ValidateNested()
+  @Type(() => AddressUpdateDto)
+  applicationAddress: AddressUpdateDto | null
+
+  @Expose()
+  @IsOptional()
+  @ValidateNested()
+  @Type(() => AddressUpdateDto)
+  applicationPickUpAddress: AddressUpdateDto | null
+
+  @Expose()
+  @IsOptional()
+  @ValidateNested()
+  @Type(() => AddressUpdateDto)
+  leasingAgentAddress: AddressUpdateDto | null
 }
 
 export class ListingUpdateDto extends ListingCreateDto {
