+++ backend/core/src/property/property.dto.ts
@@ -3,11 +3,13 @@ import { Exclude, Expose, Type } from "class-transformer"
 import { IsDefined, IsString, IsUUID, ValidateNested } from "class-validator"
 import { Property } from "../entity/property.entity"
 import { UnitDto } from "../units/unit.dto"
+import { AddressUpdateDto } from "../shared/dto/address.dto"
 
 export class PropertyDto extends OmitType(Property, [
   "listings",
   "units",
   "propertyGroups",
+  "buildingAddress"
 ] as const) {
   @Exclude()
   @ApiHideProperty()
@@ -22,6 +24,11 @@ export class PropertyDto extends OmitType(Property, [
   @ValidateNested({ each: true })
   @Type(() => UnitDto)
   units: UnitDto[]
+
+  @Expose()
+  @IsDefined()
+  @ValidateNested()
+  buildingAddress: AddressUpdateDto
 }
 
 export class PropertyCreateDto extends OmitType(PropertyDto, [
